+++ crates/tsz-checker/src/types/utilities/tests/jsdoc_params_tests.rs
@@ -855,6 +855,7 @@ const obj = {
 /// Before the fix, `assignment_anchor_node` walked up from the Parameter to
 /// the `ArrowFunction` and returned the function's start position (`(`), which
 /// shifted the diagnostic column one to the left of tsc's anchor.
+#[allow(clippy::doc_markdown)]
 #[test]
 fn jsdoc_type_function_param_default_anchors_at_parameter_name() {
     let diags = crate::test_utils::check_js_source_diagnostics(
