+++ crates/tsz-parser/src/parser/mod.rs
@@ -19,6 +19,7 @@ pub mod base;
 pub mod flags;
 pub mod node;
 mod node_access;
+mod node_access_typed_getters;
 mod node_arena;
 mod node_children;
 mod node_modifiers;
@@ -30,6 +31,7 @@ mod state_declarations;
 mod state_declarations_exports;
 mod state_expressions;
 mod state_expressions_literals;
+mod state_expressions_literals_regex;
 mod state_import_attributes;
 mod state_statements;
 mod state_statements_class;
